Move integration test helpers to integration/internal
