+++ integration/bundle/empty_bundle_test.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/databricks/cli/internal/acc"
+	"github.com/databricks/cli/integration/internal/acc"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
 )
