+++ integration/cmd/fs/helpers_test.go
@@ -5,7 +5,7 @@ import (
 	"path"
 	"path/filepath"
 
-	"github.com/databricks/cli/internal/acc"
+	"github.com/databricks/cli/integration/internal/acc"
 	"github.com/databricks/cli/internal/testutil"
 
 	"github.com/databricks/cli/libs/filer"
