+++ integration/cmd/storage_credentials/storage_credentials_test.go
@@ -3,7 +3,7 @@ package storage_credentials_test
 import (
 	"testing"
 
-	"github.com/databricks/cli/internal/acc"
+	"github.com/databricks/cli/integration/internal/acc"
 	"github.com/databricks/cli/internal/testcli"
 	"github.com/databricks/cli/internal/testutil"
 	"github.com/stretchr/testify/assert"
